stage1/requirements.txt
web3>=6.0.0
python-dotenv>=1.0.0
requests>=2.31.0
eth-abi>=4.0.0
